# Backend/swagger.yaml
swagger: "2.0"
info:
  title: Bank Argent API documentation
  description: Contains all available API endpoints in this codebase
  version: "1.0.0"
  termsOfService: "http://swagger.io/terms/"
host: localhost:3001
basePath: /api/v1
schemes:
  - http
paths:
...
  /transactions/{transactionId}:
    get:
      security:
        - Bearer: []
      tags:
        - Transactions Module
      summary: Get transaction details
      parameters:
        - name: transactionId
          in: path
          required: true
          type: string
      responses:
        "200":
          description: Transaction details
          schema:
            $ref: "#/definitions/TransactionResponse"
    put:
      security:
        - Bearer: []
      tags:
        - Transactions Module
      summary: Update a transaction
      parameters:
        - name: transactionId
          in: path
          required: true
          type: string
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/UpdateTransaction"
      responses:
        "200":
          description: Transaction updated successfully
          schema:
            $ref: "#/definitions/TransactionResponse"
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
definitions:
...
  UpdateTransaction:
    type: object
    properties:
      category: { type: string, example: "Food" }
      note: { type: string, example: "Dinner with friends" }
  Transaction:
    type: object
    properties:
      id: { type: string }
      date: { type: string, format: date-time }
      description: { type: string }
      amount: { type: number }
      type: { type: string, enum: [credit, debit] }
      category: { type: string }
      note: { type: string }
...
  TransactionResponse:
    type: object
    properties:
      status: { type: integer }
      message: { type: string }
      body:
        $ref: "#/definitions/Transaction"